requirements.txt
# Core automation
google-generativeai
anthropic
openai
youtube-transcript-api==0.6.2
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
schedule
requests
groq
flask
gunicorn>=21.2.0

# VIDEO GENERATION (Render optimized - MoviePy 1.0.3 REQUIRED)
moviepy>=2.0.0
pillow>=10.0.0
numpy>=1.24.0,<2.0.0
imageio>=2.5.0
decorator>=4.3.0
tqdm>=4.0.0
yt-dlp>=2023.10.0
edge-tts>=6.1.9
gtts>=2.3.0

# CLOUD STORAGE (REQUIRED for Render)
cloudinary>=1.36.0

# ImageMagick for moviepy text (installed via Aptfile)
# ffmpeg (installed via Aptfile)
